Project-01-Simple-QA-App/requirements.txt
streamlit
langchain
langchain-openai
langchain_community
python-dotenv
huggingface_hub
